backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
sqlparse==0.5.5
websockets==16.0
